Update README.md to reflect correct file location of csv file
while following the instructions, i experienced a fatal error with the '500_names_and_emails.csv' present in the scripts folder when trying to run 'seed_demo_data'.  script errored, insisting the csv file was not in the right location. After running generate_fake_names_and_emails, a new 500 names file was created at the root of the folder, and I was able to generate the demo data.